pyproject.toml
[project]
name = "financial-analysis"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.124.4,<0.125.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "black (>=25.12.0,<26.0.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
